+++ packages/batch-submitter/src/batch-submitter/batch-submitter.ts
@@ -3,11 +3,11 @@ import { Contract, Signer, utils, providers } from 'ethers'
 import { TransactionReceipt } from '@ethersproject/abstract-provider'
 import { Gauge, Histogram, Counter } from 'prom-client'
 import * as ynatm from '@eth-optimism/ynatm'
-import { RollupInfo } from '@eth-optimism/core-utils'
+import { RollupInfo, sleep } from '@eth-optimism/core-utils'
 import { Logger, Metrics } from '@eth-optimism/common-ts'
 import { getContractFactory } from '@metis.io/contracts'
 
-export interface Range {
+export interface BlockRange {
   start: number
   end: number
 }
@@ -79,7 +79,7 @@ export abstract class BatchSubmitter {
     endBlock: number
   ): Promise<TransactionReceipt>
   public abstract _onSync(): Promise<TransactionReceipt>
-  public abstract _getBatchStartAndEnd(): Promise<Range>
+  public abstract _getBatchStartAndEnd(): Promise<BlockRange>
   public abstract _updateChainInfo(): Promise<void>
 
   public async submitNextBatch(): Promise<TransactionReceipt> {
@@ -87,7 +87,11 @@ export abstract class BatchSubmitter {
       this.l2ChainId = await this._getL2ChainId()
     }
     await this._updateChainInfo()
-    await this._checkBalance()
+
+    if (!(await this._hasEnoughETHToCoverGasCosts())) {
+      await sleep(this.resubmissionTimeout)
+      return
+    }
 
     //this.logger.info('Readying to submit next batch...', {
     //  l2ChainId: this.l2ChainId,
@@ -108,16 +112,17 @@ export abstract class BatchSubmitter {
     return this._submitBatch(range.start, range.end)
   }
 
-  protected async _checkBalance(): Promise<void> {
+  protected async _hasEnoughETHToCoverGasCosts(): Promise<boolean> {
     const address = await this.signer.getAddress()
     const balance = await this.signer.getBalance()
     const ether = utils.formatEther(balance)
     const num = parseFloat(ether)
 
-    //this.logger.info('Checked balance', {
-    //  address,
-    //  ether,
-    //})
+    this.logger.info('Checked balance', {
+      address,
+      ether,
+    })
+
     this.metrics.batchSubmitterETHBalance.set(num)
 
     if (num < this.minBalanceEther) {
@@ -125,7 +130,10 @@ export abstract class BatchSubmitter {
         current: num,
         safeBalance: this.minBalanceEther,
       })
+      return false
     }
+
+    return true
   }
 
   protected async _getRollupInfo(): Promise<RollupInfo> {
@@ -195,7 +203,7 @@ export abstract class BatchSubmitter {
     this.metrics.batchSizeInBytes.observe(batchSizeInBytes)
     return true
   }
-  
+
   public static async getReceiptWithResubmission(
     txFunc: (gasPrice) => Promise<TransactionReceipt>,
     resubmissionConfig: ResubmissionConfig,
@@ -239,8 +247,8 @@ export abstract class BatchSubmitter {
     return minGasPriceInGwei
   }
 
- 
-  
+
+
   protected async _submitAndLogTx(
     txFunc: (gasPrice) => Promise<TransactionReceipt>,
     successMessage: string