+++ packages/batch-submitter/src/batch-submitter/state-batch-submitter.ts
@@ -12,7 +12,7 @@ import {
 import { Logger, Metrics } from '@eth-optimism/common-ts'
 
 /* Internal Imports */
-import { Range, BatchSubmitter } from '.'
+import { BlockRange, BatchSubmitter } from '.'
 
 export class StateBatchSubmitter extends BatchSubmitter {
   // TODO: Change this so that we calculate start = scc.totalElements() and end = ctc.totalElements()!
@@ -116,8 +116,8 @@ export class StateBatchSubmitter extends BatchSubmitter {
     return
   }
 
-  public async _getBatchStartAndEnd(): Promise<Range> {
-    //this.logger.info('Getting batch start and end for state batch submitter...')
+  public async _getBatchStartAndEnd(): Promise<BlockRange> {
+    this.logger.info('Getting batch start and end for state batch submitter...')
     const startBlock: number =
       (await this.chainContract.getTotalElementsByChainId(this.l2ChainId)).toNumber() +
       this.blockOffset