+++ packages/core-utils/src/watcher.ts
@@ -10,16 +10,21 @@ export interface Layer {
 export interface WatcherOptions {
   l1: Layer
   l2: Layer
+  pollInterval?: number
 }
 
 export class Watcher {
   public l1: Layer
   public l2: Layer
-  public NUM_BLOCKS_TO_FETCH: number = 10_000_000
+  public pollInterval = 3000
+  public NUM_BLOCKS_TO_FETCH = 10_000_000
 
   constructor(opts: WatcherOptions) {
     this.l1 = opts.l1
     this.l2 = opts.l2
+    if (opts.pollInterval) {
+      this.pollInterval = opts.pollInterval
+    }
   }
 
   public async getMessageHashesFromL1Tx(l1TxHash: string): Promise<string[]> {
@@ -31,14 +36,14 @@ export class Watcher {
 
   public async getL1TransactionReceipt(
     l2ToL1MsgHash: string,
-    pollForPending: boolean = true
+    pollForPending = true
   ): Promise<TransactionReceipt> {
     return this.getTransactionReceipt(this.l1, l2ToL1MsgHash, pollForPending)
   }
 
   public async getL2TransactionReceipt(
     l1ToL2MsgHash: string,
-    pollForPending: boolean = true
+    pollForPending = true
   ): Promise<TransactionReceipt> {
     return this.getTransactionReceipt(this.l2, l1ToL2MsgHash, pollForPending)
   }
@@ -71,7 +76,7 @@ export class Watcher {
   public async getTransactionReceipt(
     layer: Layer,
     msgHash: string,
-    pollForPending: boolean = true
+    pollForPending = true
   ): Promise<TransactionReceipt> {
     let matches: ethers.providers.Log[] = []
 
@@ -82,21 +87,25 @@ export class Watcher {
       const successFilter: ethers.providers.Filter = {
         address: layer.messengerAddress,
         topics: [ethers.utils.id(`RelayedMessage(bytes32)`)],
-        fromBlock: startingBlock
+        fromBlock: startingBlock,
       }
       const failureFilter: ethers.providers.Filter = {
         address: layer.messengerAddress,
         topics: [ethers.utils.id(`FailedRelayedMessage(bytes32)`)],
-        fromBlock: startingBlock
+        fromBlock: startingBlock,
       }
       const successLogs = await layer.provider.getLogs(successFilter)
       const failureLogs = await layer.provider.getLogs(failureFilter)
       const logs = successLogs.concat(failureLogs)
       matches = logs.filter((log: ethers.providers.Log) => log.data === msgHash)
+
       // exit loop after first iteration if not polling
       if (!pollForPending) {
         break
       }
+
+      // pause awhile before trying again
+      await new Promise((r) => setTimeout(r, this.pollInterval))
     }
 
     // Message was relayed in the past