+++ packages/data-transport-layer/src/services/l1-ingestion/service.ts
@@ -1,10 +1,10 @@
 /* Imports: External */
-import { fromHexString, EventArgsAddressSet } from '@eth-optimism/core-utils'
+import { fromHexString } from '@eth-optimism/core-utils'
 import { BaseService, Metrics } from '@eth-optimism/common-ts'
 import { JsonRpcProvider } from '@ethersproject/providers'
 import { LevelUp } from 'levelup'
 import { ethers, constants } from 'ethers'
-import { Gauge } from 'prom-client'
+import { Gauge, Counter } from 'prom-client'
 
 /* Imports: Internal */
 import { TransportDB, TransportDBMapHolder, TransportDBMap} from '../../db/transport-db'
@@ -24,6 +24,8 @@ import { MissingElementError, EventName } from './handlers/errors'
 
 interface L1IngestionMetrics {
   highestSyncedL1Block: Gauge<string>
+  missingElementCount: Counter<string>
+  unhandledErrorCount: Counter<string>
 }
 
 const registerMetrics = ({
@@ -33,6 +35,16 @@ const registerMetrics = ({
   highestSyncedL1Block: new client.Gauge({
     name: 'data_transport_layer_highest_synced_l1_block',
     help: 'Highest Synced L1 Block Number',
+    registers: [registry],
+  }),
+  missingElementCount: new client.Counter({
+    name: 'data_transport_layer_missing_element_count',
+    help: 'Number of times recovery from missing elements happens',
+    registers: [registry],
+  }),
+  unhandledErrorCount: new client.Counter({
+    name: 'data_transport_layer_l1_unhandled_error_count',
+    help: 'Number of times recovered from unhandled errors',
     registers: [registry],
   }),
 })
@@ -157,7 +169,8 @@ export class L1IngestionService extends BaseService<L1IngestionServiceOptions> {
 
     // Store the total number of submitted transactions so the server can tell clients if we're
     // done syncing or not
-    const totalElements = await this.state.contracts.OVM_CanonicalTransactionChain.getTotalElements()
+    const totalElements =
+      await this.state.contracts.OVM_CanonicalTransactionChain.getTotalElements()
     if (totalElements > 0) {
       await this.state.db.putHighestL2BlockNumber(totalElements - 1)
     }
@@ -229,11 +242,20 @@ export class L1IngestionService extends BaseService<L1IngestionServiceOptions> {
         }
       } catch (err) {
         if (err instanceof MissingElementError) {
+          this.logger.warn('recovering from a missing event', {
+            message: err.toString(),
+          })
+
           // Different functions for getting the last good element depending on the event type.
           const handlers = {
-            SequencerBatchAppended: this.state.db.getLatestTransactionBatch,
-            StateBatchAppended: this.state.db.getLatestStateRootBatch,
-            TransactionEnqueued: this.state.db.getLatestEnqueue,
+            SequencerBatchAppended:
+              this.state.db.getLatestTransactionBatch.bind(this.state.db),
+            StateBatchAppended: this.state.db.getLatestStateRootBatch.bind(
+              this.state.db
+            ),
+            TransactionEnqueued: this.state.db.getLatestEnqueue.bind(
+              this.state.db
+            ),
           }
 
           // Find the last good element and reset the highest synced L1 block to go back to the
@@ -267,11 +289,14 @@ export class L1IngestionService extends BaseService<L1IngestionServiceOptions> {
           )
 
           // Something we should be keeping track of.
-          this.logger.warn('recovering from a missing event', {
+          this.logger.warn('recovered from a missing event', {
             eventName,
             lastGoodBlockNumber: lastGoodElement.blockNumber,
           })
+
+          this.l1IngestionMetrics.missingElementCount.inc()
         } else if (!this.running || this.options.dangerouslyCatchAllErrors) {
+          this.l1IngestionMetrics.unhandledErrorCount.inc()
           this.logger.error('Caught an unhandled error', {
             message: err.toString(),
             stack: err.stack,
@@ -307,11 +332,14 @@ export class L1IngestionService extends BaseService<L1IngestionServiceOptions> {
     // We need to figure out how to make this work without Infura. Mark and I think that infura is
     // doing some indexing of events beyond Geth's native capabilities, meaning some event logic
     // will only work on Infura and not on a local geth instance. Not great.
-    const addressSetEvents = await this.state.contracts.Lib_AddressManager.queryFilter(
-      this.state.contracts.Lib_AddressManager.filters.AddressSet(contractName),
-      fromL1Block,
-      toL1Block
-    )
+    const addressSetEvents =
+      await this.state.contracts.Lib_AddressManager.queryFilter(
+        this.state.contracts.Lib_AddressManager.filters.AddressSet(
+          contractName
+        ),
+        fromL1Block,
+        toL1Block
+      )
 
     // We're going to parse things out in ranges because the address of a given contract may have
     // changed in the range provided by the user.