+++ packages/data-transport-layer/src/services/server/service.ts
@@ -52,7 +52,18 @@ const optionSettings = {
       return validators.isUrl(val) || validators.isJsonRpcProvider(val)
     },
   },
+  l2RpcProvider: {
+    validate: (val: unknown) => {
+      return validators.isUrl(val) || validators.isJsonRpcProvider(val)
+    },
+  },
   defaultBackend: {
+    default: 'l1',
+    validate: (val: string) => {
+      return val === 'l1' || val === 'l2'
+    },
+  },
+  l1GasPriceBackend: {
     default: 'l1',
     validate: (val: string) => {
       return val === 'l1' || val === 'l2'
@@ -70,6 +81,7 @@ export class L1TransportServer extends BaseService<L1TransportServerOptions> {
     server: any
     db: TransportDB
     l1RpcProvider: JsonRpcProvider
+    l2RpcProvider: JsonRpcProvider
   } = {} as any
 
   protected async _init(): Promise<void> {
@@ -82,6 +94,11 @@ export class L1TransportServer extends BaseService<L1TransportServerOptions> {
       typeof this.options.l1RpcProvider === 'string'
         ? new JsonRpcProvider(this.options.l1RpcProvider)
         : this.options.l1RpcProvider
+
+    this.state.l2RpcProvider =
+      typeof this.options.l2RpcProvider === 'string'
+        ? new JsonRpcProvider(this.options.l2RpcProvider)
+        : this.options.l2RpcProvider
 
     this._initializeApp()
   }
@@ -133,6 +150,26 @@ export class L1TransportServer extends BaseService<L1TransportServerOptions> {
     // and before other error middleware
     if (this.options.useSentry) {
       this.state.app.use(Sentry.Handlers.errorHandler())
+    }
+
+    this.logger.info('HTTP Server Options', {
+      defaultBackend: this.options.defaultBackend,
+      l1GasPriceBackend: this.options.l1GasPriceBackend,
+    })
+
+    if (this.state.l1RpcProvider) {
+      this.logger.info('HTTP Server L1 RPC Provider initialized', {
+        url: this.state.l1RpcProvider.connection.url,
+      })
+    } else {
+      this.logger.warn('HTTP Server L1 RPC Provider not initialized')
+    }
+    if (this.state.l2RpcProvider) {
+      this.logger.info('HTTP Server L2 RPC Provider initialized', {
+        url: this.state.l2RpcProvider.connection.url,
+      })
+    } else {
+      this.logger.warn('HTTP Server L2 RPC Provider not initialized')
     }
   }
 
@@ -241,7 +278,8 @@ export class L1TransportServer extends BaseService<L1TransportServerOptions> {
             highestL2BlockNumber = await db.getHighestL2BlockNumber()
             break
           case 'l2':
-            currentL2Block = await db.getLatestUnconfirmedTransaction()
+            currentL2Block =
+              await this.state.db.getLatestUnconfirmedTransaction()
             highestL2BlockNumber =
               (await db.getHighestSyncedUnconfirmedBlock()) - 1
             break
@@ -282,8 +320,21 @@ export class L1TransportServer extends BaseService<L1TransportServerOptions> {
     this._registerRoute(
       'get',
       '/eth/gasprice',
-      async (): Promise<GasPriceResponse> => {
-        const gasPrice = await this.state.l1RpcProvider.getGasPrice()
+      async (req): Promise<GasPriceResponse> => {
+        const backend = req.query.backend || this.options.l1GasPriceBackend
+        let gasPrice: BigNumber
+
+        if (backend === 'l1') {
+          gasPrice = await this.state.l1RpcProvider.getGasPrice()
+        } else if (backend === 'l2') {
+          const response = await this.state.l2RpcProvider.send(
+            'rollup_gasPrices',
+            []
+          )
+          gasPrice = BigNumber.from(response.l1GasPrice)
+        } else {
+          throw new Error(`Unknown L1 gas price backend: ${backend}`)
+        }
 
         return {
           gasPrice: gasPrice.toString(),
@@ -496,11 +547,12 @@ export class L1TransportServer extends BaseService<L1TransportServerOptions> {
           }
         }
 
-        const transactions = await db.getFullTransactionsByIndexRange(
-          BigNumber.from(batch.prevTotalElements).toNumber(),
-          BigNumber.from(batch.prevTotalElements).toNumber() +
-            BigNumber.from(batch.size).toNumber()
-        )
+        const transactions =
+          await this.state.db.getFullTransactionsByIndexRange(
+            BigNumber.from(batch.prevTotalElements).toNumber(),
+            BigNumber.from(batch.prevTotalElements).toNumber() +
+              BigNumber.from(batch.size).toNumber()
+          )
 
         return {
           batch,
@@ -526,11 +578,12 @@ export class L1TransportServer extends BaseService<L1TransportServerOptions> {
           }
         }
 
-        const transactions = await db.getFullTransactionsByIndexRange(
-          BigNumber.from(batch.prevTotalElements).toNumber(),
-          BigNumber.from(batch.prevTotalElements).toNumber() +
-            BigNumber.from(batch.size).toNumber()
-        )
+        const transactions =
+          await this.state.db.getFullTransactionsByIndexRange(
+            BigNumber.from(batch.prevTotalElements).toNumber(),
+            BigNumber.from(batch.prevTotalElements).toNumber() +
+              BigNumber.from(batch.size).toNumber()
+          )
 
         return {
           batch,