+++ packages/core-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @eth-optimism/core-utils
 
+## 0.5.0
+
+### Minor Changes
+
+- 049200f4: removed unused functions from core-utils
+
+## 0.4.7
+
+### Patch Changes
+
+- 224b04c0: Adds a pollInterval delay to watcher.ts
+
 ## 0.4.6
 
 ### Patch Changes